Fix error in tutorial
I was trying to figure out how inverse bind pose works, and I was really confused until I realized there's an error here compared to the actual data in "Simple Skin".

Feel free to edit this change to be prettier, but I'd like to bring it up so it can be fixed.  Perhaps it will require reediting the image as well.